+++ modules/encoding/utils.py
@@ -13,10 +13,7 @@ from configs import args, configs
 BERT_PATH       = f"{args.init_path}/_pretrained/BERT/{args.bert_model}/"
 BERT_TOKENIZER  = f"{args.init_path}/_pretrained/BERT/{args.bert_model}-vocab.txt"
 
-try:
-    BERT_tokenizer = BertTokenizer.from_pretrained(BERT_PATH, local_files_only=False)
-except OSError:
-    BERT_tokenizer = BertTokenizer.from_pretrained(args.bert_model)
+BERT_tokenizer  = BertTokenizer.from_pretrained(args.bert_model)
 nlp             = spacy.load("en_core_web_sm")